pyproject.toml
[project]
name = "pwbt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "backtrader>=1.9.78.123",
    "jupyter>=1.1.1",
    "keras>=3.8.0",
    "numpy>=2.0.2",
    "pandas>=2.2.3",
    "pq>=1.9.1",
    "psycopg>=3.2.4",
    "psycopg2-binary>=2.9.10",
    "pwb-toolbox>=0.1.1",
    "python-dotenv>=1.0.1",
    "requests-html>=0.10.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.38",
    "tensorflow-metal>=1.2.0",
    "tensorflow>=2.18.0",
    "yahoo-fin>=0.8.9.1",
]
